WEOS-40 Fix for endpoint intermittently not working as expected
The issue it seems was related to the order in which the routes were loaded.
Sorted the paths from the config alphabetically (dynamic routes have a "{"
which will be sorted after the alphas)

Also made an update where the mock handler will only be added to the endpoint
if there is an example response.